requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiohttp-retry==2.9.1
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
distro==1.9.0
fastapi==0.119.0
frozenlist==1.8.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.79
langchain-openai==0.3.35
langchain-pinecone==0.2.12
langchain-text-splitters==0.3.11
langsmith==0.4.35
MarkupSafe==3.0.3
multidict==6.7.0
numpy==2.3.3
openai==2.3.0
orjson==3.11.3
packaging==24.2
pinecone==7.3.0
pinecone-plugin-assistant==1.8.0
pinecone-plugin-interface==0.0.7
propcache==0.4.1
pydantic==2.12.2
pydantic-settings==2.11.0
pydantic_core==2.41.4
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
simsimd==6.5.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.12.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
websockets==15.0.1
yarl==1.22.0
zstandard==0.25.0
